feat(AccessFileMenu): move to files section
