+++ src/sections/file/file-action-buttons/access-file-menu/AccessStatus.tsx
@@ -1,9 +1,9 @@
-import { FilePreview } from '../../../../../../../../files/domain/models/FilePreview'
+import { FilePreview } from '../../../../files/domain/models/FilePreview'
 import { Globe, LockFill, UnlockFill } from 'react-bootstrap-icons'
 import { useTranslation } from 'react-i18next'
 import styles from './AccessFileMenu.module.scss'
 import { DropdownButtonItem } from '@iqss/dataverse-design-system'
-import { useFileDownloadPermission } from '../../../../../../../file/file-permissions/useFileDownloadPermission'
+import { useFileDownloadPermission } from '../../file-permissions/useFileDownloadPermission'
 
 interface AccessStatusProps {
   file: FilePreview
