+++ src/sections/file/file-action-buttons/access-file-menu/FileDownloadOptions.tsx
@@ -1,10 +1,10 @@
 import { DropdownHeader } from '@iqss/dataverse-design-system'
 import { Download } from 'react-bootstrap-icons'
-import { FilePreview } from '../../../../../../../../files/domain/models/FilePreview'
+import { FilePreview } from '../../../../files/domain/models/FilePreview'
 import { FileTabularDownloadOptions } from './FileTabularDownloadOptions'
 import { FileNonTabularDownloadOptions } from './FileNonTabularDownloadOptions'
 import { useTranslation } from 'react-i18next'
-import { useFileDownloadPermission } from '../../../../../../../file/file-permissions/useFileDownloadPermission'
+import { useFileDownloadPermission } from '../../file-permissions/useFileDownloadPermission'
 
 interface FileDownloadOptionsProps {
   file: FilePreview
