+++ src/sections/file/file-action-buttons/access-file-menu/FileNonTabularDownloadOptions.tsx
@@ -1,9 +1,9 @@
-import { FilePreview } from '../../../../../../../../files/domain/models/FilePreview'
-import FileTypeToFriendlyTypeMap from '../../../../../../../../files/domain/models/FileTypeToFriendlyTypeMap'
+import { FilePreview } from '../../../../files/domain/models/FilePreview'
+import FileTypeToFriendlyTypeMap from '../../../../files/domain/models/FileTypeToFriendlyTypeMap'
 import { DropdownButtonItem } from '@iqss/dataverse-design-system'
-import { useDataset } from '../../../../../../DatasetContext'
+import { useDataset } from '../../../dataset/DatasetContext'
 import { useTranslation } from 'react-i18next'
-import { FileIngestStatus } from '../../../../../../../../files/domain/models/FileIngest'
+import { FileIngestStatus } from '../../../../files/domain/models/FileIngest'
 
 interface FileNonTabularDownloadOptionsProps {
   file: FilePreview
