--- src/stories/dataset/dataset-files/files-table/file-actions/file-action-buttons/access-file-menu/RequestAccessModal.stories.tsx
@@ -1,24 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react'
-import { WithI18next } from '../../../../../../WithI18next'
-import { WithSettings } from '../../../../../../WithSettings'
-import { RequestAccessModal } from '../../../../../../../sections/dataset/dataset-files/files-table/file-actions/file-actions-cell/file-action-buttons/access-file-menu/RequestAccessModal'
-import { WithLoggedInUser } from '../../../../../../WithLoggedInUser'
-
-const meta: Meta<typeof RequestAccessModal> = {
-  title:
-    'Sections/Dataset Page/DatasetFiles/FilesTable/FileActionsCell/FileActionButtons/RequestAccessModal',
-  component: RequestAccessModal,
-  decorators: [WithI18next, WithSettings]
-}
-
-export default meta
-type Story = StoryObj<typeof RequestAccessModal>
-
-export const LoggedOut: Story = {
-  render: () => <RequestAccessModal fileId={1} />
-}
-
-export const LoggedIn: Story = {
-  decorators: [WithLoggedInUser],
-  render: () => <RequestAccessModal fileId={1} />
-}
